requirements.txt
APScheduler
discord.py
emoji
pytz
six
tzlocal
websockets
